refactor(Validator): 자동차 이름 검증과 횟수 검증 로직 분리
자동차 이름 파싱 검증과 횟수 관련 검증을 분리한다.
