+++ src/main/java/racingcar/utils/Parser.java
@@ -4,12 +4,12 @@ import static racingcar.constant.SymbolType.CAR_NAME_DELIMITER;
 
 import java.util.Arrays;
 import java.util.List;
-import racingcar.validator.Validator;
+import racingcar.validator.ParserValidator;
 
 public final class Parser {
     public static List<String> parse(String input) {
         List<String> names = Arrays.stream(input.split(CAR_NAME_DELIMITER.getSymbol())).map(String::trim).toList();
-        Validator.namesValidate(names);
+        ParserValidator.namesValidate(names);
         return names;
     }
 }
